Fixed the days being off by 1 after noon...
Since this is a count down to NOON -- ( or any time passed with hours and minutes to an exact hour ) the day would be still say the number of days once the hour past. IE: If noon was the time at 1130 it would have said 75 days, 0 hours, .... then after noon, it would say 75 days, 23 hours - rather than 74 days and 23 hours.

I also dropped seconds for now until I get it worked out - it does not sync with minutes, reason because I want hole numbers but milliseconds is a decimal.